# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scrap Payments Manager API",
        "description": "API documentation for Scrap Payments Manager application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/credits": {
            "get": {
                "tags": [
                    "Credits"
                ],
                "summary": "Listar todos los créditos",
                "operationId": "06524fc31f65d27206387f3f86bdb87e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "Filtrar por cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Credits"
                ],
                "summary": "Crear un nuevo crédito",
                "operationId": "54d3ad1ffab7a7860f5167f2c9b62975",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clientId",
                                    "date",
                                    "items"
                                ],
                                "properties": {
                                    "clientId": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-24"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "description": {
                                                    "type": "string",
                                                    "example": "Material de construcción"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recurso creado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/credits/{id}": {
            "get": {
                "tags": [
                    "Credits"
                ],
                "summary": "Obtener detalles de un crédito específico",
                "operationId": "ee1d81f706fc9fc241c362c075b50ec7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del crédito",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Credits"
                ],
                "summary": "Actualizar un crédito existente",
                "operationId": "3b4699dce73e08145d983070295a2849",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del crédito",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "description": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado exitosamente"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clients/{clientId}/credits": {
            "get": {
                "tags": [
                    "Credits"
                ],
                "summary": "Obtener todos los créditos de un cliente específico",
                "operationId": "ee380413f834f949f1736785d9704d4d",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}